Add The Battle of Five Armies rules.
This is the "Advanced Rules" lists, with the exception that including a
second (or more) Goblin Shaman in place of a Chieftain is not implemented.

Wizards or Heroes can be Generals, so this is done using an upgrade.

Links to the rules are [available here](https://www.reddit.com/r/warmaster/comments/tblgxh/battle_of_five_armies/).